fix some logic for answer question

+++ src/bot/handlers/contestant_question_handler.py
@@ -1,3 +1,5 @@
+from pyexpat.errors import messages
+
 from aiogram import Router, F, types, Bot
 from aiogram.fsm.context import FSMContext
 from aiogram.types import Message
@@ -5,9 +7,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from bot.callbacks.contestant_question_factory import ContestantQuestionCallbackFactory
 from bot.enums import QuestionState
-from bot.handlers.contestant_handler import callback_back
 from bot.handlers.question_handler import print_profile
-from bot.keyboards.contestant_question_kb import question_keyboard, question_error_keyboard
+from bot.keyboards.contestant_question_kb import  question_error_keyboard
 from bot.states import StatesBot
 from database.crud.questions import update_state, get_question, add_answer_to_db
 
@@ -26,16 +27,17 @@ async def waiting_response_callback(
     await callback.message.delete()
     data = await state.get_data()
     question_id = callback_data.question_id
+    question = await get_question(question_id=question_id, db_session=db_session)
+
     msg = await bot.send_message(
         chat_id=chat_id,
-        text="Введите ответ на вопрос текстом:",
-        reply_markup=question_keyboard(question_id=question_id),
+        text=f"Введите ответ на вопрос \"{question.question}\"текстом:",
+        reply_markup=question_error_keyboard(question_id=question_id),
     )
-    messages = data.get("message_for_delete", [])
-    messages.append(msg.message_id)
-    question = await get_question(question_id=question_id, db_session=db_session)
+    messages_list = data.get("message_for_delete", [])
+    messages_list.append(msg.message_id)
     await state.update_data(
-        message_for_delete=messages,
+        message_for_delete=messages_list,
         contestant_id=question.contestant_id,
         user_id=question.user_id,
         question_id=question_id,
@@ -43,15 +45,23 @@ async def waiting_response_callback(
     await state.set_state(StatesBot.ANSWER_QUESTION)
 
 
-# Перейти в состоянеи ожидания сообщения
 @router.callback_query(ContestantQuestionCallbackFactory.filter(F.state == QuestionState.REJECTED))
 async def reject_callback(
     callback: types.CallbackQuery,
     callback_data: ContestantQuestionCallbackFactory,
     db_session: AsyncSession,
     state: FSMContext,
+    bot: Bot,
 ):
-    await callback.message.delete()
+    data = await state.get_data()
+    messages_list = data.get("message_for_delete", [])
+    if not messages_list:
+        await callback.message.delete()
+    await state.update_data(message_for_delete=[])
+    for msg in messages_list:
+        await bot.delete_message(chat_id=callback.message.chat.id, message_id=msg)
+
+
     await update_state(question_id=callback_data.question_id, state=QuestionState.REJECTED, db_session=db_session)
     await callback.answer("Вопрос отклонён")
     await state.clear()
@@ -65,10 +75,10 @@ async def get_any_message(message: Message, state: FSMContext):
         "Вы ввели не текст! Ответ должен состоять из текста. Введите ответ еще раз:",
         reply_markup=question_error_keyboard(question_id=question_id),
     )
-    messages = data.get("message_for_delete", [])
-    messages.append(msg.message_id)
-    messages.append(message.message_id)
-    await state.update_data(message_for_delete=messages)
+    messages_list = data.get("message_for_delete", [])
+    messages_list.append(msg.message_id)
+    messages_list.append(message.message_id)
+    await state.update_data(message_for_delete=messages_list)
     return
 
 
@@ -88,9 +98,9 @@ async def get_message(message: Message, state: FSMContext, db_session: AsyncSess
     by_msg = await message.answer("Спасибо за ответ!")
 
     # delete messages
-    messages = data.get("message_for_delete", [])
+    messages_list = data.get("message_for_delete", [])
     await state.update_data(message_for_delete=[])
-    for msg in messages:
+    for msg in messages_list:
         await bot.delete_message(chat_id=by_msg.chat_id, message_id=msg)
 
     await bot.delete_message(chat_id=by_msg.chat_id, message_id=by_msg.message_id)
